+++ src/lib/scaffold-templater.ts
@@ -6,6 +6,7 @@
 import * as fs from 'fs-extra';
 import * as path from 'path';
 import * as _ from 'lodash';
+import * as detectIndent from 'detect-indent';
 
 import { FileDataNode, FileDataNodeType, Scaffold } from '../contracts/scaffold';
 import { TemplateString } from './template-string';
@@ -66,6 +67,12 @@ export module ScaffoldTemplater {
             } else {
                 let templateContent = fs.readFileSync(node.fullPath).toString();
                 let outputContent = (options.header || '') + _.template(templateContent)(data);
+
+                if (options.indention) {
+                    const { indent } = detectIndent(outputContent);
+                    outputContent = outputContent.replace(new RegExp(indent, 'g'), options.indention);
+                }
+
                 fs.writeFileSync(targetFullPath, outputContent);
             }
         }