+++ test/data/scaffolds/Example/___.uniq(stuff).join('-')__.example.ts
@@ -1,1 +1,8 @@
-console.log('this is example.ts - <%= name.toCamelCase() %>');
+function <%= name.toCamelCase() %>(returnYes: boolean): string {
+    console.log('this is example.ts - <%= name.toCamelCase() %>');
+
+    if (returnYes) {
+        return 'YESS!!!'
+    }
+    return '...umm no';
+}