+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Change Log
 
+## 0.2.0
+
+* Enable `indention` field in scaffs-config (override template's indention)
+
 ## 0.1.0
 
-* Enable `header` field in scaffs-config (prepended to every file)
+* Enable `header` field in scaffs-config (text prepended to every file)
 
 ## 0.0.2 - 0.0.7